feat: port core library modules (types, formats, yaml)
