--- src/types/index.ts
@@ -1,5 +1,0 @@
-/**
- * @module types
- */
-
-export * from "@lodestar/types";
